refactor: split rumor add/delete commands into rumors_admin.go
